pref: add ststus validation; update db table

+++ src/tests/test_users.py
@@ -50,11 +50,12 @@ def test_get_user():
 
 
         response = client.get(f"/users/{user_id}")
-        if response:
-            new_user_id = data['res']['id']
-            assert new_user_id == user_id
+        assert response.status_code == 200, response.text
+        new_data = response.json()
+        assert new_data['res']['id'] == user_id
     finally:
-        db.delete_user(user_id)
+        if user_id is not None:
+            db.delete_user(user_id)
 
 def test_update_users():
     user_id = None
@@ -66,7 +67,7 @@ def test_update_users():
             "phone": "phone1",
             "email": "mail",
         })
-        assert response.status_code == 201, response.text
+        assert response.status_code == 200, response.text
 
         data = response.json()
         user_id = data["res"]["id"]
@@ -78,6 +79,7 @@ def test_update_users():
         response = client.patch(f"/users/{user_id}", json = {
             "nickname": "nameTest"
         })
+        assert response.status_code == 201, response.text
 
         saved_user = db.get_user(user_id)
         assert "nameTest" == saved_user["nickname"]
